saving column names in system namespace
