+++ src/storage/src/frontend/tests/mod.rs
@@ -1,0 +1,42 @@
+use super::*;
+
+#[cfg(test)]
+mod queries;
+#[cfg(test)]
+mod schema;
+// Copyright 2020 Alex Dukhno
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#[cfg(test)]
+mod table;
+
+fn create_table<P: backend::BackendStorage>(
+    storage: &mut FrontendStorage<P>,
+    schema_name: &str,
+    table_name: &str,
+    column_names: Vec<&str>,
+) {
+    storage
+        .create_schema(schema_name)
+        .expect("no system errors")
+        .expect("schema is created");
+    storage
+        .create_table(
+            schema_name,
+            table_name,
+            column_names.into_iter().map(ToOwned::to_owned).collect::<Vec<String>>(),
+        )
+        .expect("no system errors")
+        .expect("table is created");
+}
